split bundles for hooks and icons
