+++ packages/core/helper-plugin/lib/src/components/FilterListURLQuery/FilterListURLQuery.js
@@ -6,7 +6,7 @@
 
 import React from 'react';
 import PropTypes from 'prop-types';
-import useQueryParams from '../../hooks/useQueryParams';
+import { useQueryParams } from '../../hooks/useQueryParams';
 import AttributeTag from './AttributeTag';
 
 const FilterListURLQuery = ({ filtersSchema }) => {
