+++ packages/core/helper-plugin/lib/src/components/PaginationURLQuery/PaginationURLQuery.js
@@ -20,7 +20,7 @@ import PropTypes from 'prop-types';
 import { useLocation } from 'react-router-dom';
 import { useIntl } from 'react-intl';
 import { stringify } from 'qs';
-import useQueryParams from '../../hooks/useQueryParams';
+import { useQueryParams } from '../../hooks/useQueryParams';
 
 const PaginationURLQuery = ({ pagination: { pageCount } }) => {
   const [{ query }] = useQueryParams();
