+++ packages/core/helper-plugin/lib/src/hooks/useAppInfos/index.js
@@ -1,16 +1,1 @@
-/**
- *
- * useAppInfos
- *
- */
-
-import { useContext } from 'react';
-import { AppInfosContext } from '../../contexts/AppInfosContext';
-
-const useAppInfos = () => {
-  const appInfos = useContext(AppInfosContext);
-
-  return appInfos;
-};
-
-export default useAppInfos;
+export * from './useAppInfos';
