+++ packages/core/helper-plugin/lib/src/hooks/useAppInfos/useAppInfos.js
@@ -1,0 +1,16 @@
+/**
+ *
+ * useAppInfos
+ *
+ */
+
+import { useContext } from 'react';
+import { AppInfosContext } from '../../contexts/AppInfosContext';
+
+const useAppInfos = () => {
+  const appInfos = useContext(AppInfosContext);
+
+  return appInfos;
+};
+
+export { useAppInfos };
