--- packages/core/helper-plugin/lib/src/hooks/useFormattedMessage/index.js
@@ -1,17 +1,0 @@
-import { useIntl } from 'react-intl';
-import isObject from 'lodash/isObject';
-
-const useFormattedMessage = message => {
-  const { formatMessage } = useIntl();
-
-  if (isObject(message) && message.id) {
-    return formatMessage({
-      ...message,
-      defaultMessage: message.defaultMessage || message.id,
-    });
-  }
-
-  return message;
-};
-
-export default useFormattedMessage;
