+++ packages/core/helper-plugin/lib/src/hooks/useLockScroll/useLockScroll.js
@@ -1,0 +1,15 @@
+import { useEffect } from 'react';
+
+const useLockScroll = lockScroll => {
+  useEffect(() => {
+    if (lockScroll) {
+      document.body.classList.add('lock-body-scroll');
+    }
+
+    return () => {
+      document.body.classList.remove('lock-body-scroll');
+    };
+  }, [lockScroll]);
+};
+
+export { useLockScroll };
