+++ packages/core/helper-plugin/lib/src/hooks/useNotification/index.js
@@ -1,19 +1,1 @@
-/**
- *
- * useNotification
- *
- */
-
-import { useContext, useRef } from 'react';
-import { NotificationsContext } from '../../contexts/NotificationsContext';
-
-const useNotification = () => {
-  const { toggleNotification } = useContext(NotificationsContext);
-  // Use a ref so we can safely add the toggleNotification
-  // to a hook dependencies array
-  const toggleNotificationRef = useRef(toggleNotification);
-
-  return toggleNotificationRef.current;
-};
-
-export default useNotification;
+export * from './useNotification';
