+++ packages/core/helper-plugin/lib/src/hooks/usePersistentState/usePersistentState.js
@@ -1,0 +1,17 @@
+import { useState, useEffect } from 'react';
+
+const usePersistentState = (key, defaultValue) => {
+  const [value, setValue] = useState(() => {
+    const stickyValue = window.localStorage.getItem(key);
+
+    return stickyValue !== null ? JSON.parse(stickyValue) : defaultValue;
+  });
+
+  useEffect(() => {
+    window.localStorage.setItem(key, JSON.stringify(value));
+  }, [key, value]);
+
+  return [value, setValue];
+};
+
+export { usePersistentState };
