+++ packages/core/helper-plugin/lib/src/hooks/useQuery/index.js
@@ -1,7 +1,1 @@
-import { useLocation } from 'react-router-dom';
-
-const useQuery = () => {
-  return new URLSearchParams(useLocation().search);
-};
-
-export default useQuery;
+export * from './useQuery';
