+++ packages/core/helper-plugin/lib/src/hooks/useQuery/useQuery.js
@@ -1,0 +1,7 @@
+import { useLocation } from 'react-router-dom';
+
+const useQuery = () => {
+  return new URLSearchParams(useLocation().search);
+};
+
+export { useQuery };
