+++ packages/core/helper-plugin/lib/src/hooks/useRBACProvider/index.js
@@ -1,12 +1,1 @@
-/**
- *
- * useRBACProvider
- *
- */
-
-import { useContext } from 'react';
-import { RBACProviderContext } from '../../contexts/RBACProviderContext';
-
-const useRBACProvider = () => useContext(RBACProviderContext);
-
-export default useRBACProvider;
+export * from './useRBACProvider';
