+++ packages/core/helper-plugin/lib/src/hooks/useRBACProvider/useRBACProvider.js
@@ -1,0 +1,12 @@
+/**
+ *
+ * useRBACProvider
+ *
+ */
+
+import { useContext } from 'react';
+import { RBACProviderContext } from '../../contexts/RBACProviderContext';
+
+const useRBACProvider = () => useContext(RBACProviderContext);
+
+export { useRBACProvider };
