+++ packages/core/helper-plugin/lib/src/hooks/useStrapiApp/index.js
@@ -1,12 +1,1 @@
-/**
- *
- * useStrapiApp
- *
- */
-
-import { useContext } from 'react';
-import { StrapiAppContext } from '../../contexts/StrapiAppContext';
-
-const useStrapiApp = () => useContext(StrapiAppContext);
-
-export default useStrapiApp;
+export * from './useStrapiApp';
