+++ packages/core/helper-plugin/lib/src/hooks/useStrapiApp/useStrapiApp.js
@@ -1,0 +1,12 @@
+/**
+ *
+ * useStrapiApp
+ *
+ */
+
+import { useContext } from 'react';
+import { StrapiAppContext } from '../../contexts/StrapiAppContext';
+
+const useStrapiApp = () => useContext(StrapiAppContext);
+
+export { useStrapiApp };
